+++ newclient/src/modules/address/handlers/AddressOverview/AddressOverview.vue
@@ -9,7 +9,7 @@ import { Component, Mixins, Prop, Watch } from 'vue-property-decorator'
 import { exchangeRate } from '@app/modules/addresses/addresses.graphql'
 import { FormattedNumber } from '@app/core/helper/number-format-helper'
 import { NumberFormatMixin } from '@app/core/components/mixins/number-format.mixin'
-import { getEthBalance, getContractMeta } from './addressDetails.graphql'
+import { getEthBalance, getContractMeta, getContractTimestamp } from './addressDetails.graphql'
 import { getEthBalance_getEthBalance as BalanceType } from './apolloTypes/getEthBalance'
 import { getContractMeta_getContractMeta as ContractMeta } from './apolloTypes/getContractMeta'
 import { Address } from '@app/modules/address/components/props'
@@ -52,7 +52,8 @@ import { ErrorMessage } from '@app/modules/address/models/ErrorMessagesForAddres
             result(data) {
                 if (data.data && data.data.getContractMeta) {
                     this.emitErrorState(false)
-                    this.setContract(true)
+                    this.setContract(true, data.data.getContractMeta)
+                    this.getTimestamp(data.data.getContractMeta)
                 } else {
                     this.setContract(false)
                 }
@@ -79,12 +80,13 @@ export default class AddressOverview extends Mixins(CoinData) {
 
     @Prop(String) address!: string
     @Prop(Boolean) isMiner!: boolean
-    @Prop(Boolean) isContractcreator!: boolean
+    @Prop(Boolean) isContractCreator!: boolean
     @Prop(Boolean) isContract!: boolean
     @Prop(Number) totalErc20Owned!: number
     @Prop(Boolean) loadingTokens!: boolean
     @Prop(Boolean) updateBalance!: boolean
     @Prop(Function) setContract!: void
+    @Prop(Function) setContractTimestamp!: (data: number) => void
     /*
     ===================================================================================
       Initial Data
@@ -105,7 +107,7 @@ export default class AddressOverview extends Mixins(CoinData) {
             hash: this.address,
             balance: this.getEthBalance ? this.getEthBalance.balance : '0',
             isMiner: this.isMiner,
-            isContractCreator: this.isContractcreator,
+            isContractCreator: this.isContractCreator,
             totalERC20: this.totalErc20Owned,
             isContract: this.isContract
         }
@@ -119,6 +121,24 @@ export default class AddressOverview extends Mixins(CoinData) {
       Methods
     ===================================================================================
     */
+    getTimestamp(contract: ContractMeta) {
+        this.$apollo
+            .query({
+                query: getContractTimestamp,
+                variables: {
+                    hash: contract.transactionHash
+                }
+            })
+            .then(response => {
+                if (response.data) {
+                    this.setContractTimestamp(response.data.getTransactionByHash.timestamp)
+                }
+            })
+            .catch(error => {
+                this.emitErrorState(true, ErrorMessage.contractTimestampNotFound)
+            })
+    }
+
     /**
      * Emits error to Sentry
      */