+++ newclient/src/modules/address/pages/PageDetailsAddress.vue
@@ -14,12 +14,13 @@
                     <address-overview
                         :address="addressRef"
                         :is-miner="isMiner"
-                        :is-contractcreator="isContractCreator"
+                        :is-contract-creator="isContractCreator"
                         :is-contract="isContract"
                         :total-erc20-owned="totalERC20"
                         :update-balance="addrBalanceChanged"
                         :set-contract-creator="setContractCreator"
                         :set-contract="setContract"
+                        :set-contract-timestamp="setContractTimestamp"
                         :loading-tokens="loadingERC20Balance"
                         @resetBalanceUpdate="resetBalance"
                         @errorAddrOverview="setError"
@@ -57,6 +58,7 @@
                                     v-show="toggleLastTx === 0"
                                     :address="addressRef"
                                     :max-items="max"
+                                    :is-contract="isContract"
                                     :new-transfers="newETHTransfers"
                                     @resetUpdateCount="setNewEvent"
                                     @errorTransfers="setError"
@@ -176,7 +178,7 @@
                   MINING HISTORY INFO TAB
                 =====================================================================================
                 -->
-                <v-tab-item slot="tabs-item" value="tab-3">
+                <v-tab-item v-if="showMiningRewards" slot="tabs-item" value="tab-3">
                     <keep-alive>
                         <div>
                             <v-layout :class="subMenuClass" row wrap align-center justify-start pt-2>
@@ -270,8 +272,14 @@
                   CONTRACT CREATOR INFO TAB
                 =====================================================================================
                 -->
-                <v-tab-item v-if="isContractCreator" slot="tabs-item" value="tab-4">
-                    <!-- <table-address-contracts v-if="activeTab === 'tab-4'" :address="addressRef" /> -->
+                <v-tab-item v-if="isContract" slot="tabs-item" :value="showMiningRewards ? 'tab-4' : 'tab-3'">
+                    <keep-alive>
+                        <v-layout row wrap justify-start class="mb-4 contract-layout">
+                            <v-flex xs12>
+                                <app-details-list :is-loading="loadingContractDetails" :title="$t('contract.details')" :details="details" />
+                            </v-flex>
+                        </v-layout>
+                    </keep-alive>
                 </v-tab-item>
             </app-tabs>
         </div>
@@ -285,7 +293,7 @@ import AppMessage from '@app/core/components/ui/AppMessage.vue'
 import AppTabs from '@app/core/components/ui/AppTabs.vue'
 import { AddressUpdateEvent } from '@app/modules/address/handlers/AddressUpdateEvent/AddressUpdateEvent.mixin'
 import { Component, Prop, Mixins, Watch } from 'vue-property-decorator'
-import { Crumb, Tab } from '@app/core/components/props'
+import { Detail, Crumb, Tab } from '@app/core/components/props'
 import AppInfoLoad from '@app/core/components/ui/AppInfoLoad.vue'
 import { eth } from '@app/core/helper'
 import AddressPendingTx from '@app/modules/address/handlers/AddressPendingTx/AddressPendingTx.vue'
@@ -293,8 +301,9 @@ import AddressOverview from '@app/modules/address/handlers/AddressOverview/Addre
 import AddressTransfers from '@app/modules/address/handlers/AddressTransfers/AddressTransfers.vue'
 import AddressTokens from '@app/modules/address/handlers/AddressTokens/AddressTokens.vue'
 import AddressRewards from '@app/modules/address/handlers/AddressRewards/AddressRewards.vue'
-import { Address } from '@app/modules/address/components/props'
+import { Address, Contract } from '@app/modules/address/components/props'
 import { ErrorMessage } from '@app/modules/address/models/ErrorMessagesForAddress'
+import AppDetailsList from '@app/core/components/ui/AppDetailsList.vue'
 
 const MAX_ITEMS = 10
 
@@ -309,7 +318,8 @@ const MAX_ITEMS = 10
         AddressTransfers,
         AddressTokens,
         AddressRewards,
-        AddressPendingTx
+        AddressPendingTx,
+        AppDetailsList
     }
 })
 export default class PageDetailsAddress extends Mixins(AddressUpdateEvent) {
@@ -326,6 +336,7 @@ export default class PageDetailsAddress extends Mixins(AddressUpdateEvent) {
       Initial Data
     ===================================================================================
     */
+    loadingContractDetails = true
     isContractCreator = false
     isContract = false
     error = ''
@@ -338,6 +349,7 @@ export default class PageDetailsAddress extends Mixins(AddressUpdateEvent) {
     hasUncleRewards = false
     hasBlockRewards = false
     errorMessages: ErrorMessage[] = []
+    contract!: Contract
 
     /* ERC20 and ERC721 Refetc options */
     totalERC20 = 0
@@ -348,6 +360,53 @@ export default class PageDetailsAddress extends Mixins(AddressUpdateEvent) {
       Computed Values
     ===================================================================================
     */
+    get details(): Detail[] {
+        let details: Detail[] = []
+        if (this.loadingContractDetails) {
+            details = [
+                {
+                    title: this.$i18n.t('contract.date-created')
+                },
+                {
+                    title: this.$i18n.t('contract.tx-hash')
+                },
+                {
+                    title: this.$i18n.t('contract.creator')
+                },
+                {
+                    title: this.$i18n.t('contract.code-hash')
+                }
+            ]
+        } else {
+            details = [
+                {
+                    title: this.$i18n.t('contract.date-created'),
+                    detail: this.timestamp
+                },
+                {
+                    title: this.$i18n.t('contract.tx-hash'),
+                    detail: this.contract.transactionHash,
+                    link: `/tx/${this.contract.transactionHash}`,
+                    copy: true,
+                    mono: true
+                },
+                {
+                    title: this.$i18n.t('contract.creator'),
+                    detail: this.contract.creator,
+                    link: `/address/${this.contract.creator}`,
+                    copy: true,
+                    mono: true
+                },
+                {
+                    title: this.$i18n.t('contract.code-hash'),
+                    detail: this.contract.codeHash,
+                    copy: true,
+                    mono: true
+                }
+            ]
+        }
+        return details
+    }
 
     get isValid(): boolean {
         return eth.isValidAddress(this.addressRef)
@@ -399,10 +458,10 @@ export default class PageDetailsAddress extends Mixins(AddressUpdateEvent) {
                 tabs.push(newTab)
             }
 
-            if (this.isContractCreator) {
+            if (this.isContract) {
                 const newTab = {
-                    id: 4,
-                    title: this.$i18n.tc('contract.name', 2).toString(),
+                    id: this.showMiningRewards ? 4 : 3,
+                    title: this.$i18n.tc('contract.name', 1).toString(),
                     isActive: false
                 }
                 tabs.push(newTab)
@@ -421,12 +480,18 @@ export default class PageDetailsAddress extends Mixins(AddressUpdateEvent) {
     get subMenuClass(): string {
         return this.$vuetify.breakpoint.name === 'xs' || this.$vuetify.breakpoint.name === 'sm' ? 'pt-3' : 'pt-2'
     }
+    get timestamp() {
+        return new Date(this.contract.timestamp * 1e3).toString()
+    }
 
     /*
     ===================================================================================
       Methods
     ===================================================================================
     */
+    getTimestamp() {
+        return new Date(this.contract.timestamp * 1e3).toString()
+    }
     /**
      * Sets Block Rewards
      * @param value {Boolean}
@@ -459,8 +524,17 @@ export default class PageDetailsAddress extends Mixins(AddressUpdateEvent) {
      * Sets Contract
      * @param value {Boolean}
      */
-    setContract(value: boolean): void {
+    setContract(value: boolean, data: Contract): void {
         this.isContract = value
+        this.contract = data
+    }
+    /**
+     * Sets Contract Timestamp
+     * @param timestamp {Number}
+     */
+    setContractTimestamp(timestamp: number): void {
+        this.$set(this.contract, 'timestamp', timestamp)
+        this.loadingContractDetails = false
     }
     /**
      * Sets Total Tokens
@@ -526,7 +600,7 @@ export default class PageDetailsAddress extends Mixins(AddressUpdateEvent) {
 }
 </script>
 
-<style scoped lang="css">
+<style scoped lang="scss">
 .active-button {
     color: #3d55a5;
     margin: 0px 10px;
@@ -557,5 +631,12 @@ export default class PageDetailsAddress extends Mixins(AddressUpdateEvent) {
 .tab-fade-leave-active,
 .tab-fade-leave-to {
     display: none;
+}
+.contract-layout {
+    padding: 20px;
+    .flex {
+        border: 1px solid #efefef;
+        border-radius: 4px;
+    }
 }
 </style>