+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 ### Bug 
-- Hide state diff for failed tx [#913](https://github.com/EthVM/EthVM/pull/913)
+- Query failed tx state diff if null [#913](https://github.com/EthVM/EthVM/pull/913)
+
+### Release v1.0.4
+
+### Feature
+- Sort Address Tokens [#909](https://github.com/EthVM/EthVM/pull/909)
+- Add in Address Contract Details [#914](https://github.com/EthVM/EthVM/pull/914)
+
+===================================================================================
 
 ### Release v1.0.3-hotfix.2
 
 - Fix variables [#908](https://github.com/EthVM/EthVM/pull/908)
+
+===================================================================================
 
 ### Release v1.0.3-hotfix.1